Revert "Swap highlight theme to default and default to legacy"
This reverts commit 53382e6c8d730ce483b6505c8e958a1358796da0.

The "legacy" theme should still be the default as it works well under both
dark and light background.

The "highlight" theme should be named "dark", and a new "light" theme
should be added. These 3 are the supported themes and user override can
be implemented with a local .json config file in the future.